requirements_pthon27.txt
appdirs==1.4.4
blessings==1.7
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
distlib==0.3.1
dlib==19.21.1
filelock==3.0.12
functools32==3.2.3.post2
gpustat==0.6.0
importlib-metadata==2.1.1
importlib-resources==3.3.1
matplotlib==2.0.0
numpy==1.14.5
pandas==0.7.0
pathlib2==2.3.5
Pillow==6.0.0
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==1.5
pytz==2021.1
scandir==1.10.0
scikit-image==0.10.0
scipy==1.2.0
singledispatch==3.4.0.3
six==1.15.0
subprocess32==3.5.4
torch==0.4.1
torchvision==0.1.6
typing==3.7.4.3
zipp==1.2.0
opencv-python==3.4.0.12
